Clean up use of const vs non-const XferItem objects.
To enable clean use of const in client code, the EventToUser factory
functions must be clear about the constness of the XferItem object
they are expecting as a parameter. The semantics are, if the XferItem
object passed to an EventToUser factory function as const pointer,
then the event does not take ownership of the object, and the object
is not freed when the event is freed; if the object is non-const, the
event takes ownership and the object is freed when the event is freed.

In case of the EVENT_TO_USER_REPORT_PROGRESS event, the XferItem
object must be const because it is used read-only and does not get
freed in events_to_user_free().

In case of the EVENT_TO_USER_DONE event, the object must be non-const
because it gets freed in events_to_user_free().

Change-Id: Ia44401b2ddd7c240838c919123eb097a9bf63bb2