Fix compiler warning for libcurl 7.39.0 and up.
At the time of this writing, StrBo Linux uses libcurl 7.40.0, but
Ubuntu 14.04.x ships with libcurl 7.35.0. A new error code was
introduced in 7.39.0 that was not handled in our big switch which maps
libcurl errors to list errors, yielding a compiler warning because
there is no default case (rightly so!).

An explicit compile-time version check has been added to resolve the
problem (the error codes are enumerated values, not macros).

Change-Id: Ie7230c40f34e9930d26aa5dee8e3fd059107e29f