Set maximum number of HTTP redirects to 5.
As a result of server misconfiguration, a total of 3 redirects were
suddenly required to download files from our server. The SPI slave
specifies an HTTP URL for its update XML file, which was redirected
twice by the misconfigured server to HTTPS URLs, only to be redirected
again to another HTTP URL.

Example:
http://www.ta-hifi.de/fileadmin/auto_download/MP8/MP8_UPD.xml    -> 301 Moved Permanently
https://ta-hifi.de//fileadmin/auto_download/MP8/MP8_UPD.xml      -> 301 Moved Permanently
https://www.ta-hifi.de/fileadmin/auto_download/MP8/MP8_UPD.xml   -> 301 Moved Permanently
http://updates.ta-hifi.de/firmware/auto_download/MP8/MP8_UPD.xml -> 200 OK

Even though this meant that customers were unable to perform software
updates for a few hours, RT still thinks this is good behavior
because (1) ideally, no redirects should be required, but allowing for
a single redirect should be good enough to cover all sane use cases
(server migration); our previous limit of 2 was a gratuitous choice
already; (2) the need for many redirects is a pretty clear sign for
some kind of misconfiguration which must be fixed; (3) increasing the
limit may hide server misconfigurations; and (4) a limit of 5 is still
not enough in case 6 or more redirects are required. JK thinks
differently and wants to avoid trouble on customer side.

This, however, could be achieved in much better ways if everybody
would just think before doing something, and test after doing
something.

Change-Id: I72cfeb35548c35b8dca4685325f4f073427f7338